+++ src/doomstat.h
@@ -516,6 +516,10 @@ extern int show_berserk;
 // Status Bar/HUD
 extern int alt_arms;
 extern int smarttotals;
+extern int hudcolor_wg_name; // [Cherry]
+extern int hudcolor_wg_name_cons; // [Cherry]
+extern int hudcolor_plain; // [Cherry]
+extern int hudcolor_totaltime; // [Cherry]
 extern int hudcolor_ms_incomp;
 extern int hudcolor_ms_comp;
 typedef enum {